@@ -8,12 +8,12 @@
     "not OperaMini all"
   ],
   "scripts": {
-    "build": "./build",
-    "build-dev": "SKYNET_MYSKY_ENV=dev ./build",
-    "build-alpha": "./build",
+    "build": "NODE_ENV=production npm run build:css && ./build",
+    "build-dev": "npm run build:css && SKYNET_MYSKY_ENV=dev ./build",
+    "build:css": "postcss assets/style.compile.css -o assets/style.css",
     "deploy": "npm run build && ts-node deploy.ts",
     "deploy-dev": "npm run build-dev && DEV=1 ts-node deploy.ts",
-    "deploy-alpha": "npm run build-alpha && ALPHA=1 ts-node deploy.ts",
+    "deploy-alpha": "npm run build && ALPHA=1 ts-node deploy.ts",
     "lint": "npm run lint:tsc && npm run lint:eslint",
     "lint:eslint": "eslint --ext .ts src scripts tests deploy.ts",
     "lint:tsc": "tsc --noemit && tsc --project tsconfig.scripts.json --noemit",
@@ -32,21 +32,26 @@
   "homepage": "https://github.com/SkynetLabs/skynet-mysky#readme",
   "lint-staged": {
     "*.{js,jsx,ts,tsx}": [
-      "eslint --ext .ts --max-warnings 0",
       "prettier --write"
     ],
     "*.{json,yml,md}": [
       "prettier --write"
     ]
   },
+  "jest": {
+    "testEnvironment": "jsdom"
+  },
   "dependencies": {
     "buffer": "^6.0.3",
     "confusables": "^1.0.0",
-    "idb-keyval": "^5.1.5",
+    "crypto-browserify": "^3.12.0",
+    "idb-keyval": "^6.0.2",
+    "mustache": "^4.2.0",
     "post-me": "^0.4.5",
     "punycode": "^2.1.1",
     "skynet-js": "^4.0.12-beta",
     "skynet-mysky-utils": "^0.3.0",
+    "stream-browserify": "^3.0.0",
     "tweetnacl": "^1.0.3",
     "url-join": "^4.0.1"
   },
@@ -55,23 +60,29 @@
     "@babel/preset-env": "^7.15.6",
     "@babel/preset-typescript": "^7.15.0",
     "@skynetlabs/skynet-nodejs": "^2.2.0",
+    "@tailwindcss/forms": "^0.3.3",
     "@types/jest": "^27.0.1",
+    "@types/mustache": "^4.1.2",
     "@types/url-join": "^4.0.1",
     "@typescript-eslint/eslint-plugin": "^4.31.0",
     "@typescript-eslint/parser": "^4.31.0",
+    "autoprefixer": "^10.3.4",
     "chalk": "^4.1.2",
     "eslint": "^7.32.0",
     "eslint-plugin-jsdoc": "^36.1.0",
     "husky": "^7.0.2",
     "ifdef-loader": "^2.1.5",
-    "jest": "^26.6.3",
+    "jest": "^27.2.0",
     "jsdom": "^17.0.0",
     "lint-staged": "^11.1.2",
-    "prettier": "^2.4.0",
+    "postcss": "^8.3.6",
+    "postcss-cli": "^8.2.0",
+    "prettier": "^2.4.1",
     "rimraf": "^3.0.2",
+    "tailwindcss": "^2.2.15",
     "ts-loader": "^9.2.5",
     "typescript": "^4.4.3",
-    "webpack": "^5.52.1",
+    "webpack": "^5.53.0",
     "webpack-cli": "^4.8.0"
   }
 }
